requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib[bcrypt]
bcrypt
python-jose
groq
pydantic
fastapi-mail
celery
redis
jinja2
